move 02_12 functions into HCFA_Info class, delete 02_12.inc.php
